pyproject.toml
[project]
name = "dedicatorias"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.31",
    "openai>=1.12",
    "google-genai>=0.5",
    "pillow>=10.0",
    "reportlab>=4.0",
    "pypdfium2>=4.0",
    "boto3>=1.34",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]
